+++ 1.20/Common/src/main/java/fuzs/puzzleslib/impl/client/init/ItemDisplayOverridesImpl.java
@@ -4,7 +4,6 @@ import com.google.common.base.Preconditions;
 import com.google.common.collect.Maps;
 import com.google.common.collect.Sets;
 import fuzs.puzzleslib.api.client.init.v1.ItemModelDisplayOverrides;
-import fuzs.puzzleslib.impl.client.core.ClientFactories;
 import net.minecraft.client.resources.model.BakedModel;
 import net.minecraft.client.resources.model.ModelResourceLocation;
 import net.minecraft.world.item.ItemDisplayContext;
@@ -15,8 +14,6 @@ import java.util.Map;
 import java.util.Objects;
 
 public abstract class ItemDisplayOverridesImpl implements ItemModelDisplayOverrides {
-    public static final ItemDisplayOverridesImpl INSTANCE = ClientFactories.INSTANCE.getItemModelDisplayOverrides();
-
     protected final Map<ModelResourceLocation, Map<ItemDisplayContext, ModelResourceLocation>> overrideLocations = Maps.newHashMap();
 
     @Override