+++ 1.20/Fabric/src/main/java/fuzs/puzzleslib/impl/core/FabricFactories.java
@@ -5,10 +5,12 @@ import fuzs.puzzleslib.api.core.v1.ModConstructor;
 import fuzs.puzzleslib.api.init.v2.GameRulesFactory;
 import fuzs.puzzleslib.api.init.v2.PotionBrewingRegistry;
 import fuzs.puzzleslib.api.item.v2.ToolTypeHelper;
+import fuzs.puzzleslib.api.item.v2.crafting.CombinedIngredients;
 import fuzs.puzzleslib.impl.event.FabricEventInvokerRegistryImpl;
 import fuzs.puzzleslib.impl.init.FabricGameRulesFactory;
 import fuzs.puzzleslib.impl.init.PotionBrewingRegistryFabric;
 import fuzs.puzzleslib.impl.item.FabricToolTypeHelper;
+import fuzs.puzzleslib.impl.item.crafting.FabricCombinedIngredients;
 
 import java.util.Set;
 
@@ -52,5 +54,10 @@ public final class FabricFactories implements CommonFactories {
     @Override
     public ToolTypeHelper getToolTypeHelper() {
         return new FabricToolTypeHelper();
+    }
+
+    @Override
+    public CombinedIngredients getCombinedIngredients() {
+        return new FabricCombinedIngredients();
     }
 }