+++ 1.20/Fabric/src/main/java/fuzs/puzzleslib/impl/event/GrindstoneXpHolder.java
@@ -1,0 +1,6 @@
+package fuzs.puzzleslib.impl.event;
+
+public interface GrindstoneXpHolder {
+
+    int puzzleslib$getXp();
+}