+++ 1.20/Forge/src/main/java/fuzs/puzzleslib/impl/core/BiomeLoadingHandler.java
@@ -6,21 +6,34 @@ import com.mojang.serialization.Codec;
 import fuzs.puzzleslib.api.biome.v1.BiomeLoadingContext;
 import fuzs.puzzleslib.api.biome.v1.BiomeLoadingPhase;
 import fuzs.puzzleslib.api.biome.v1.BiomeModificationContext;
+import fuzs.puzzleslib.api.core.v1.ContentRegistrationFlags;
+import fuzs.puzzleslib.api.resources.v1.AbstractModPackResources;
+import fuzs.puzzleslib.api.resources.v1.PackResourcesHelper;
 import fuzs.puzzleslib.impl.biome.*;
 import net.minecraft.core.Holder;
+import net.minecraft.network.chat.CommonComponents;
+import net.minecraft.network.chat.Component;
+import net.minecraft.resources.ResourceLocation;
+import net.minecraft.server.packs.PackType;
+import net.minecraft.server.packs.repository.RepositorySource;
+import net.minecraft.world.flag.FeatureFlagSet;
 import net.minecraft.world.level.biome.Biome;
 import net.minecraftforge.common.world.BiomeModifier;
 import net.minecraftforge.common.world.ModifiableBiomeInfo;
 import net.minecraftforge.eventbus.api.IEventBus;
 import net.minecraftforge.registries.DeferredRegister;
 import net.minecraftforge.registries.ForgeRegistries;
+import org.apache.commons.lang3.StringUtils;
 import org.jetbrains.annotations.Nullable;
 
-import java.util.Collection;
-import java.util.HashMap;
-import java.util.Map;
+import java.io.ByteArrayInputStream;
+import java.nio.charset.StandardCharsets;
+import java.util.*;
 import java.util.function.Consumer;
+import java.util.function.Function;
 import java.util.function.Predicate;
+import java.util.stream.Collectors;
+import java.util.stream.Stream;
 
 /**
  * Forge implementation based on <a href="https://github.com/teamfusion/rottencreatures/blob/1.19.2/forge/src/main/java/com/github/teamfusion/platform/common/worldgen/forge/BiomeManagerImpl.java">BiomeManager</a>
@@ -34,11 +47,31 @@ public class BiomeLoadingHandler {
         this.put(BiomeModifier.Phase.MODIFY, BiomeLoadingPhase.MODIFICATIONS);
         this.put(BiomeModifier.Phase.AFTER_EVERYTHING, BiomeLoadingPhase.POST_PROCESSING);
     }});
+    private static final String BIOME_MODIFICATIONS_NAME_KEY = "biome_modifications";
+    private static final String BIOME_MODIFIERS_DATA_KEY = ForgeRegistries.Keys.BIOME_MODIFIERS.location().toString().replace(":", "/");
+    private static final Function<String, ResourceLocation> BIOME_MODIFICATIONS_FILE_KEY = id -> new ResourceLocation(id, BIOME_MODIFIERS_DATA_KEY + "/" + id + ".json");
+    private static final Function<ResourceLocation, String> BIOME_MODIFICATIONS_FILE_CONTENTS = id -> "{\"type\":\"" + id + "\"}";
 
     public static void register(String modId, IEventBus modEventBus, Multimap<BiomeLoadingPhase, BiomeModification> biomeModifications) {
         DeferredRegister<Codec<? extends BiomeModifier>> deferredRegister = DeferredRegister.create(ForgeRegistries.Keys.BIOME_MODIFIER_SERIALIZERS, modId);
         deferredRegister.register(modEventBus);
-        deferredRegister.register("biome_modifications", new BiomeModifierImpl(biomeModifications)::codec);
+        deferredRegister.register(BIOME_MODIFICATIONS_NAME_KEY, new BiomeModifierImpl(biomeModifications)::codec);
+    }
+
+    public static RepositorySource buildPack(String modId) {
+        ResourceLocation id = new ResourceLocation(modId, BiomeLoadingHandler.BIOME_MODIFICATIONS_NAME_KEY);
+        String title = Stream.of(ContentRegistrationFlags.BIOME_MODIFICATIONS.name().toLowerCase(Locale.ROOT).split("_")).map(StringUtils::capitalize).collect(Collectors.joining(" "));
+        return PackResourcesHelper.buildServerPack(() -> new AbstractModPackResources() {
+
+            @Override
+            public void listResources(PackType packType, String namespace, String path, ResourceOutput resourceOutput) {
+                if (path.equals(BiomeLoadingHandler.BIOME_MODIFIERS_DATA_KEY)) {
+                    resourceOutput.accept(BiomeLoadingHandler.BIOME_MODIFICATIONS_FILE_KEY.apply(modId), () -> {
+                        return new ByteArrayInputStream(BiomeLoadingHandler.BIOME_MODIFICATIONS_FILE_CONTENTS.apply(id).getBytes(StandardCharsets.UTF_8));
+                    });
+                }
+            }
+        }, id, Component.literal(title), CommonComponents.EMPTY, true, true, FeatureFlagSet.of(), true);
     }
 
     private record BiomeModifierImpl(Multimap<BiomeLoadingPhase, BiomeModification> biomeModifications,